+++ src/main/java/com/codeactuator/rocket/dao/WorkforceRepository.java
@@ -5,5 +5,5 @@ package com.codeactuator.rocket.dao;
 import com.codeactuator.rocket.domain.Workforce;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WorkforceRepository extends CrudRepository<Workforce, Integer> {
+public interface WorkforceRepository extends CrudRepository<Workforce, Long> {
 }
